refactor: extract utils and cleanup comments
- Extract getLocaleStartEnd to src/utils/dateUtils.js
- Extract calculateStreak to src/utils/streakUtils.js
- Update habit service to use extracted utils
- Remove non-JSDoc comments from all project JS files
- Add missing JSDocs
